Moved a function around
