+++ test/http_client.c
@@ -2,7 +2,6 @@
 /*
  * http_client.c -- A simple HTTP/QUIC client
  */
-
 #ifndef WIN32
 #include <arpa/inet.h>
 #include <netinet/in.h>
@@ -28,6 +27,7 @@
 #endif
 #include <sys/stat.h>
 #include <fcntl.h>
+#include <event2/event.h>
 
 #include "lsquic.h"
 #include "test_common.h"
@@ -159,6 +159,65 @@ struct lsquic_stream_ctx {
     struct lsquic_reader reader;
 };
 
+
+/*Function resolves a Hostname into an Ip Adress
+Parameters:
+-hostname	the URL of the website
+-version	0 for ipv4 and 1 for ipv6
+Partly taken from https://www.binarytides.com/hostname-to-ip-address-c-sockets-linux/ */
+int get_Ip_from_DNS(const char* hostname, char* ipaddr, int version, const char* port)
+{
+    struct addrinfo hints, *servinfo, *p;
+    struct sockaddr_in *h;
+    struct sockaddr_in6 *h6;
+    int rv;
+    char * port2 = strdup(port); /*so that it's no longer const */
+    char * hostname2 = strdup(hostname);
+
+    memset(&hints, 0, sizeof(hints));
+    if (version)
+    {
+        hints.ai_family = AF_INET6;
+    }
+    else
+    {
+        hints.ai_family = AF_INET;
+    }
+
+    hints.ai_socktype = SOCK_STREAM;
+
+
+    if ((rv = getaddrinfo(hostname2, port2, &hints, &servinfo)) != 0)
+    {
+        LSQ_ERROR("getaddrinfo: %s\n", gai_strerror(rv));
+        free(port2);
+        free(hostname2);
+        freeaddrinfo(servinfo);
+        return 1;
+    }
+
+    if (version)/*Ipv6*/
+    {
+        for (p = servinfo; p != NULL; p = p->ai_next)
+        {
+            h6 = (struct sockaddr_in6*) p->ai_addr;
+            inet_ntop(AF_INET6, &h6->sin6_addr, ipaddr, 47);
+        }
+    }
+    else /*Ipv4*/
+    {
+        for (p = servinfo; p != NULL; p = p->ai_next)
+        {
+            h = (struct sockaddr_in *) p->ai_addr;
+            inet_ntop(AF_INET, &h->sin_addr, ipaddr, 47);
+        }
+    }
+
+    free(hostname2);
+    free(port2);
+    freeaddrinfo(servinfo);
+    return 0;
+}
 
 static lsquic_stream_ctx_t *
 http_client_on_new_stream (void *stream_if_ctx, lsquic_stream_t *stream)
@@ -436,7 +495,7 @@ main (int argc, char **argv)
     struct prog prog;
 
     ip = (char *)malloc(47 * sizeof(char)); /*46 is the maximal length of an ipv6 adress + '\0' character*/
-    ipv6 = false;
+    ipv6 = 0;
 
     TAILQ_INIT(&sports);
     memset(&client_ctx, 0, sizeof(client_ctx));
@@ -460,7 +519,7 @@ main (int argc, char **argv)
     {
         switch (opt) {
         case '6':
-            ipv6 = true;
+            ipv6 = 1;
             break;
         case 'I':
             client_ctx.hcc_flags |= HCC_ABORT_ON_INCOMPLETE;
