+++ test/prog.h
@@ -6,14 +6,9 @@
 #ifndef PROG_H
 #define PROG_H 1
 
-//Define Boolean type
-typedef int bool;
-#define true 1
-#define false 0
-
 
 char *ip;	/*The ipadress that will be determined. For easier access it defined here.*/
-bool ipv6;	/*True = Program uses ipv6, False = Program uses ipv4*/
+int ipv6;	/*True = Program uses ipv6, False = Program uses ipv4*/
 
 struct event;
 struct event_base;
@@ -91,9 +86,4 @@ prog_is_stopped (void);
 void
 prog_process_conns (struct prog *);
 
-/*Function resolves a Hostname into an Ip Adress
-Parameters: 
-    -hostname	the URL of the website
-    -version	0 for ipv4 and 1 for ipv6*/
-int getIpfromDNS(const char* hostname, char* ipaddr, bool version, const char* port);
 #endif
